refactor(plugin): extract report workflow helpers
